+++ osu.Game/Rulesets/RulesetStore.cs
@@ -50,7 +50,7 @@ namespace osu.Game.Rulesets
         /// </summary>
         /// <param name="id">The ruleset's internal ID.</param>
         /// <returns>A ruleset, if available, else null.</returns>
-        public RulesetInfo GetRuleset(int id) => AvailableRulesets.FirstOrDefault(r => r.ID == id);
+        public RulesetInfo GetRuleset(long id) => AvailableRulesets.FirstOrDefault(r => r.ID == id);
 
         /// <summary>
         /// Retrieve a ruleset using a known short name.
